First half of governance deployment scripts

+++ protocol/contracts/common/utils/IWETH10.sol
@@ -11,9 +11,8 @@ import "../../interfaces/token/ERC20/IERC20.sol";
 /// `withdraw` ETH from WETH10, which will then burn WETH10 token in your wallet. The amount of WETH10 token in any wallet is always identical to the
 /// balance of ETH deposited minus the ETH withdrawn with that specific wallet.
 interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
-
 	/// @dev Returns current amount of flash-minted WETH10 token.
-	function flashMinted() external view returns(uint256);
+	function flashMinted() external view returns (uint256);
 
 	/// @dev `msg.value` of ETH sent to this contract grants caller account a matching increase in WETH10 token balance.
 	/// Emits {Transfer} event to reflect WETH10 token mint of `msg.value` from `address(0)` to caller account.
@@ -42,7 +41,11 @@ interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
 	/// Requirements:
 	///   - `from` account must have at least `value` balance of WETH10 token.
 	///   - `from` account must have approved caller to spend at least `value` of WETH10 token, unless `from` and caller are the same account.
-	function withdrawFrom(address from, address payable to, uint256 value) external;
+	function withdrawFrom(
+		address from,
+		address payable to,
+		uint256 value
+	) external;
 
 	/// @dev `msg.value` of ETH sent to this contract grants `to` account a matching increase in WETH10 token balance,
 	/// after which a call is executed to an ERC677-compliant contract with the `data` parameter.
@@ -56,7 +59,11 @@ interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
 	/// Emits {Approval} event.
 	/// Returns boolean value indicating whether operation succeeded.
 	/// For more information on {approveAndCall} format, see https://github.com/ethereum/EIPs/issues/677.
-	function approveAndCall(address spender, uint256 value, bytes calldata data) external returns (bool);
+	function approveAndCall(
+		address spender,
+		uint256 value,
+		bytes calldata data
+	) external returns (bool);
 
 	/// @dev Moves `value` WETH10 token from caller's account to account (`to`),
 	/// after which a call is executed to an ERC677-compliant contract with the `data` parameter.
@@ -66,5 +73,9 @@ interface IWETH10 is IERC20, IERC2612, IERC3156FlashLender {
 	/// Requirements:
 	///   - caller account must have at least `value` WETH10 token.
 	/// For more information on {transferAndCall} format, see https://github.com/ethereum/EIPs/issues/677.
-	function transferAndCall(address to, uint value, bytes calldata data) external returns (bool);
+	function transferAndCall(
+		address to,
+		uint256 value,
+		bytes calldata data
+	) external returns (bool);
 }
